backend/requirements.txt
asgiref==3.8.1
channels-redis==4.2.0
daphne==4.1.2
Django==5.0.4
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
djangorestframework==3.15.1
psycopg2-binary>=2.9
sqlparse==0.5.0
uvicorn==0.29.0